+++ CHANGELOG.md
@@ -4,15 +4,17 @@ All notable changes to the "vscode-abap-remote-fs" extension will be documented 
 
 Format based on [Keep a Changelog](http://keepachangelog.com/)
 
-## Unreleased
+## 1.0.0
 
 ### Added
 
 - AbapGit stage
 - AbapGit push
+- Abapgit versioned source
 
 ### Changed
 
+- links in git repositories
 - adopted semantyc versioning
 
 ## [0.5.37] 2020-03-23